==> 1_Pengumpulan_Data.ipynb <==
# Pengumpulan Data
Pada penelitian analisis sentimen ini menggunakan data sebanyak 9.908 tweet dari API Twitter dengan kata kunci "anies", "prabowo", "ganjar", "muhaimin", "gibran", dan "mahfud" menggunakan library Python yang bernama tweet-harvest. 

import os
import pandas as pd
from sklearn.model_selection import train_test_split
from modul._2_preprocessing_data import normalization

from warnings import filterwarnings
filterwarnings("ignore")

folder_path = 'data/manual/'
file_list = [os.path.join(folder_path, file) for file in os.listdir(folder_path)]
display(file_list)

df = pd.concat([pd.read_excel(file, usecols=['full_text', 'sentiment']) for file in file_list], ignore_index=True)
display(df)

df['clean_full_text'] = df['full_text'].apply(lambda data: normalization(text=data, remove_punctuation_number=False))
df['word_count'] = df["clean_full_text"].str.split().str.len()
display(df)
display(df.describe())
display(df['sentiment'].value_counts())

df_train, df_test = train_test_split(df, test_size=0.2, stratify=df['sentiment'], random_state=42)
df_train = pd.DataFrame(df_train).reset_index(drop=True)
df_test = pd.DataFrame(df_test).reset_index(drop=True)

display(df_train)
display(df_train.describe())
display(df_train['sentiment'].value_counts())

display(df_test)
display(df_test.describe())
display(df_test['sentiment'].value_counts())

df_train.to_csv('output/1_pengumpulan_data/train.csv', index=False)
df_test.to_csv('output/1_pengumpulan_data/test.csv', index=False)
