==> 3_Feature_Extraction.ipynb <==
# Feature Extraction
Feature extraction adalah proses mengambil suatu data atau informasi yang berhubungan (relevan) dari data mentah atau dataset yang umumnya bersifat kompleks untuk digunakan dalam melakukan analisis atau melakukan pemodelan data (visualisasi data). (Skill, 2023a) 

import pandas as pd
from modul._2_preprocessing_data import cleaning_data
from modul._3_feature_extraction import tfidf, punctuation, pos_tagging

from warnings import filterwarnings
filterwarnings("ignore")

df_train = pd.read_csv('output/2_preprocessing_data/train.csv')
df_test = pd.read_csv('output/2_preprocessing_data/test.csv')
display(df_train)
display(df_test)

display(df_train.shape)
display(df_train['full_text'].duplicated().sum())
display(df_train.isnull().sum())
df_train = cleaning_data(df_train)
display(df_train.shape)

display(df_test.shape)
display(df_test['full_text'].duplicated().sum())
display(df_test.isnull().sum())
df_test = cleaning_data(df_test)
display(df_test.shape)

### 1. TF - IDF
Term Frequency - Invers Document Frequency (TF - IDF) adalah statistik numerik yang menunjukkan relevansi kata kunci dengan beberapa dokumen tertentu atau dapat dikatakan bahwa ia menyediakan kata kunci tersebut, dengan begitu beberapa dokumen tertentu dapat diidentifikasi atau dikategorikan. TF - IDF terdapat dua perhitungan, yaitu Term Frequency (TF) dan Invers Document Frequency (IDF). (Alrajak et al., 2020) 

tfidf_df_train = tfidf(df_train)
tfidf_df_test = tfidf(df_test)
display(tfidf_df_train)
display(tfidf_df_test)

tfidf_df_train.to_parquet('output/3_feature_extraction/train_tfidf.parquet', index=False)
tfidf_df_test.to_parquet('output/3_feature_extraction/test_tfidf.parquet', index=False)

### 2. Punctuation
Punctuation (tanda baca) adalah tanda yang dipakai dalam sistem ejaan. Tanda baca dapat membantu pembaca untuk memahami makna tulisan dengan tepat. Pada penelitian ini, punctuation akan digunakan sebagai fitur, mulai dari tanda titik (.), tanda koma (,), tanda seru (!), dan lain-lain. (Suyatno et al., 2017)

punctuation_df_train = punctuation(df_train)
punctuation_df_test = punctuation(df_test)
display(punctuation_df_train)
display(punctuation_df_test)

punctuation_df_train.to_csv('output/3_feature_extraction/train_punctuation.csv', index=False)
punctuation_df_test.to_csv('output/3_feature_extraction/test_punctuation.csv', index=False)

### 3. POS Tagging
Part-of-Speech (POS) Tagging merupakan tugas mendasar dalam Natural Language Processing (NLP) yang melibatkan penetapan kategori gramatikal (seperti kata benda, kata kerja, kata sifat, dsb.). Secara konsep, POS Tag adalah kode pendek yang mewakili bagian-bagian tertentu dari suatu ujaran. Pada penelitian ini, POS Tag akan digunakan sebagai fitur, meliputi kata benda (NN), kata kerja (VB), kata sifat (JJ), dan lain-lain. (Mudadla, 2023)

pos_tagging_df_train = pos_tagging(df_train)
pos_tagging_df_test = pos_tagging(df_test)
display(pos_tagging_df_train)
display(pos_tagging_df_test)

pos_tagging_df_train.to_csv('output/3_feature_extraction/train_pos_tagging.csv', index=False)
pos_tagging_df_test.to_csv('output/3_feature_extraction/test_pos_tagging.csv', index=False)

### Download

df_train[['full_text', 'sentiment']].to_csv('output/3_feature_extraction/train.csv', index=False)
df_test[['full_text', 'sentiment']].to_csv('output/3_feature_extraction/test.csv', index=False)
