==> 4_Feature_Engineering.ipynb <==
# Feature Engineering
Feature engineering merupakan sebuah proses untuk mengembangkan dan memilih suatu fitur atau atribut (features) yang akan digunakan untuk melakukan analisis data atau dalam melakukan pembuatan sebuah model machine learning. Pada penelitian ini, feature engineering akan digunakan untuk konversi (one-hot encoding) data kategorikal seperti fitur label sentiment dan label figure, yang dijadikan sebuah format yang dapat digunakan oleh sebuah model. (Skill, 2023b)

import pandas as pd

from warnings import filterwarnings
filterwarnings("ignore")

df_train = pd.read_csv('output/3_feature_extraction/train.csv')
df_test = pd.read_csv('output/3_feature_extraction/test.csv')
display(df_train)
display(df_test)

df_train.loc[df_train['sentiment'] == 'positif', 'sentiment'] = df_test.loc[df_test['sentiment'] == 'positif', 'sentiment'] = 1
df_train.loc[df_train['sentiment'] == 'netral', 'sentiment'] = df_test.loc[df_test['sentiment'] == 'netral', 'sentiment'] = 0
df_train.loc[df_train['sentiment'] == 'negatif', 'sentiment'] = df_test.loc[df_test['sentiment'] == 'negatif', 'sentiment'] = -1
display(df_train)
display(df_test)

tfidf_df_train = pd.read_parquet('output/3_feature_extraction/train_tfidf.parquet')
tfidf_df_test = pd.read_parquet('output/3_feature_extraction/test_tfidf.parquet')
display(tfidf_df_train)
display(tfidf_df_test)

punctuation_df_train = pd.read_csv('output/3_feature_extraction/train_punctuation.csv')
punctuation_df_test = pd.read_csv('output/3_feature_extraction/test_punctuation.csv')
display(punctuation_df_train)
display(punctuation_df_test)

pos_tagging_df_train = pd.read_csv('output/3_feature_extraction/train_pos_tagging.csv')
pos_tagging_df_test = pd.read_csv('output/3_feature_extraction/test_pos_tagging.csv')
display(pos_tagging_df_train)
display(pos_tagging_df_test)

df_train = pd.concat([tfidf_df_train, punctuation_df_train, pos_tagging_df_train, df_train[['sentiment']]], axis=1)
df_test = pd.concat([tfidf_df_test, punctuation_df_test, pos_tagging_df_test, df_test[['sentiment']]], axis=1)
display(df_train)
display(df_test)

df_train.to_parquet('output/4_feature_engineering/train.parquet', index=False)
df_test.to_parquet('output/4_feature_engineering/test.parquet', index=False)
