==> 7_Modelling.ipynb <==
# Implementasi Algoritma
Setelah melakukan dimensional reduction, data tersebut dibagi menjadi data training dan data testing dengan perbandingan 80:20. Pada pengimplementasian menggunakan kedua algoritma tersebut akan menggunakan parameter berupa data training, fitur kueri, dan nilai k. Adapun nilai k yang digunakan dalam penelitian ini sebesar 10 sampai 200 dengan interval 10. Sehingga output yang dihasilkan berupa prediksi label dari fitur kueri tersebut.

import pandas as pd
import numpy as np
from modul._7_modelling import DWKNN

from warnings import filterwarnings
filterwarnings("ignore")

df_train = pd.read_parquet('output/4_feature_engineering/train.parquet')
df_test = pd.read_parquet('output/4_feature_engineering/test.parquet')
display(df_train)
display(df_train['sentiment'].value_counts())
display(df_test)
display(df_test['sentiment'].value_counts())

# oversampling
# df_train = pd.concat([df_train[df_train['sentiment'] == 0], df_train[df_train['sentiment'] == 1].sample(n=4502, replace=True, random_state=42), df_train[df_train['sentiment'] == -1].sample(n=4502, replace=True, random_state=42)], ignore_index=True)
# df_test = pd.concat([df_test[df_test['sentiment'] == 0], df_test[df_test['sentiment'] == 1].sample(n=1181, replace=True, random_state=42), df_test[df_test['sentiment'] == -1].sample(n=1181, replace=True, random_state=42)], ignore_index=True)
# display(df_train)
# display(df_train['sentiment'].value_counts())
# display(df_test)
# display(df_test['sentiment'].value_counts())

# undersampling
# df_train = pd.concat([df_train[df_train['sentiment'] == 0].sample(n=460, replace=True, random_state=42), df_train[df_train['sentiment'] == 1].sample(n=460, replace=True, random_state=42), df_train[df_train['sentiment'] == -1]], ignore_index=True)
# df_test = pd.concat([df_test[df_test['sentiment'] == 0].sample(n=119, replace=True, random_state=42), df_test[df_test['sentiment'] == 1].sample(n=119, replace=True, random_state=42), df_test[df_test['sentiment'] == -1]], ignore_index=True)
# display(df_train)
# display(df_train['sentiment'].value_counts())
# display(df_test)
# display(df_test['sentiment'].value_counts())

X_train = df_train.drop(columns=['sentiment'])
X_test = df_test.drop(columns=['sentiment'])
y_train = df_train['sentiment']
y_test = df_test['sentiment']

X_train = X_train[list(set(X_test.columns) & set(X_train.columns))]
for col in list(set(X_test.columns) - set(X_train.columns)):
  X_train[col] = 0
X_train = X_train[sorted(X_train.columns)]
X_test = X_test[sorted(X_test.columns)]
display(X_train)
display(X_test)

X_train = np.array(X_train)
X_test = np.array(X_test)
y_train = np.array(y_train)
y_test = np.array(y_test)

def classification(k_value, X_train, y_train, X_test):
  result = pd.DataFrame()
  dwknn = DWKNN(k=k_value)
  dwknn.fit(X_train, y_train)
  y_predict = dwknn.predict(X_test)
  result[f'k_{k_value}'] = pd.Series(y_predict)
  result.to_csv(f'output/7_modelling/predict_k_{k_value:02}.csv', index=False)

df_test[['sentiment']].to_csv('output/7_modelling/actual.csv', index=False)

# classification(5, X_train, y_train, X_test)
# classification(10, X_train, y_train, X_test)
# classification(15, X_train, y_train, X_test)
# classification(20, X_train, y_train, X_test)
# classification(25, X_train, y_train, X_test)
# classification(30, X_train, y_train, X_test)
# classification(35, X_train, y_train, X_test)
# classification(40, X_train, y_train, X_test)
# classification(45, X_train, y_train, X_test)
# classification(50, X_train, y_train, X_test)
